refactor: replace old hide with new dyncamic allowcation
