--- cmd/syncthing/console_stub.go
@@ -1,17 +1,0 @@
-// Copyright (C) 2014 The Syncthing Authors.
-//
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// You can obtain one at https://mozilla.org/MPL/2.0/.
-
-//go:build !windows
-// +build !windows
-
-package main
-
-func AttachConsole() error {
-	return nil // No-op for non-Windows
-}
-func InitConsole() error {
-	return nil // No-op for non-Windows
-}
